refactor: split admin_repo.go into sub-domain files
Splits 1,275-line god-file into per-sub-domain files following the
F1-F5 boundaries already documented in AdminRepository interface:
customers, stats, billing, salepages, pixels, replays, events, audit.

Core struct, constructor, and interface unchanged. Pure file
relocation — no behavior change.
